+++ packages/prettier/src/options/sort-wechat-svg.ts
@@ -1,0 +1,56 @@
+import type { Options } from 'prettier'
+
+// https://prettier.io/docs/options
+// https://github.com/NiklasPor/prettier-plugin-organize-attributes
+export default {
+  parser: 'html',
+  printWidth: 450,
+  htmlWhitespaceSensitivity: 'ignore',
+  plugins: ['prettier-plugin-organize-attributes'],
+  attributeSort: 'ASC',
+  attributeGroups: [
+    '^xmlns$',
+    '^viewBox$',
+    '^preserveAspectRatio$',
+    '^id$',
+    '^name',
+    '^class$',
+    '^x$',
+    '^cx$',
+    '^y$',
+    '^cy$',
+    '^_?width$',
+    '^_?height$',
+    '^r$',
+    '^opacity$',
+
+    '^attributeName$',
+    '^attributeType$',
+    '^type$',
+    '^begin$',
+    '^end$',
+    '^dur$',
+    '^from$',
+    '^to$',
+    '^by$',
+    '^values$',
+    '^keyTimes$',
+    '^keySplines$',
+    '^calcMode$',
+    '^repeatCount$',
+    '^repeatDur$',
+    '^fill$',
+    '^restart$',
+    '^accumulate$',
+    '^additive$',
+
+    '^stroke*$',
+    '^style$',
+    '^src$',
+    '^href$',
+    '^d$',
+
+    '^data-*$',
+    '^version$',
+  ],
+} satisfies Options