+++ packages/prettier/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## prettier@0.0.3 (2026-02-21)
+
+### Features
+
+- Add attribute sorting options for HTML, Vue, and WXML files
+
 ## prettier@0.0.2 (2026-02-20)
 
 - Update `exports` field for CommonJS compatibility